Add CitizenNumber to urgent requests and update DTOs
- Introduced `CitizenNumber` property in `UrgentBloodRequest` and related DTOs.
- Added `BloodTypeResponseDTO` class for blood type responses.
- Updated `CreateUrgentBloodRequestDTO` to include `CitizenNumber`.
- Removed namespaces from `BloodTypeResponseDTO` and `CreateUrgentBloodRequestDTO` for code restructuring.
