--- Blood Donation Support/DTO/BloodTypeResponseDTO.cs
@@ -1,8 +1,0 @@
-namespace Blood_Donation_Support.DTO
-{
-    public class BloodTypeResponseDTO
-    {
-        public int BloodTypeId { get; set; }
-        public string BloodTypeName { get; set; } = null!;
-    }
-}
